renamed the validator file to a clearer name
